score/requirements.txt
pandas>=2.2.0
numpy>=2.0.0
scipy>=1.14.0
matplotlib>=3.8.0
seaborn>=0.13.0
pillow>=11.3.0
paho-mqtt>=2.1.0